refactor: remove unnecessary 'Data' suffix
