+++ hooks/useNodeReleases.ts
@@ -1,0 +1,4 @@
+import { useContext } from 'react';
+import { NodeReleasesContext } from '../providers/nodeReleasesProvider';
+
+export const useNodeReleases = () => useContext(NodeReleasesContext);
