--- hooks/useNodeReleasesData.ts
@@ -1,4 +1,0 @@
-import { useContext } from 'react';
-import { NodeReleasesDataContext } from '../providers/nodeReleasesDataProvider';
-
-export const useNodeReleasesData = () => useContext(NodeReleasesDataContext);
